Generate a Plain List Matrix

This function takes a listing object and converts it into a plain matrix format.

mtx_plain_list(list)

Arguments

list

A listing object to be converted into a matrix.

Value

A matrix representation of the input list.

Examples

# Example usage:
lst <- gen_plain_list("Patient", 5,
  num_cols = c("Age", "Weight", "Height"),
  cat_cols = c("Sex", "BMI"), date_cols = c("DOB", "Visit")
)
mtx_plain_list(lst)
#>      Patient    Age     Weight  Height  Sex  BMI  DOB          Visit       
#> [1,] "########" "##.##" "##.##" "##.##" "xx" "xx" "YYYY-MM-DD" "YYYY-MM-DD"
#> [2,] "########" "##.##" "##.##" "##.##" "xx" "xx" "YYYY-MM-DD" "YYYY-MM-DD"
#> [3,] "########" "##.##" "##.##" "##.##" "xx" "xx" "YYYY-MM-DD" "YYYY-MM-DD"
#> [4,] "########" "##.##" "##.##" "##.##" "xx" "xx" "YYYY-MM-DD" "YYYY-MM-DD"
#> [5,] "########" "##.##" "##.##" "##.##" "xx" "xx" "YYYY-MM-DD" "YYYY-MM-DD"